Allow to define an arbitrary service to be used
